+++ service/src/main/java/com/narel/spring/dto/CarReadDto.java
@@ -6,6 +6,8 @@ import com.narel.spring.enums.Type;
 import lombok.Value;
 import lombok.experimental.FieldNameConstants;
 
+import java.math.BigDecimal;
+import java.time.Instant;
 import java.util.List;
 
 @Value
@@ -15,7 +17,7 @@ public class CarReadDto {
     Integer id;
     Integer year;
     String image;
-    Double rentalPrice;
+    BigDecimal rentalPrice;
     String registrationNumber;
     String brand;
     String model;
